+++ src/main/java/com/epam/reportportal/cucumber/StepReporter.java
@@ -67,9 +67,10 @@ public class StepReporter extends AbstractReporter {
 
 	@Override
 	protected void beforeStep(TestStep testStep) {
-		Step step = currentScenarioContext.getStep(testStep);
+		RunningContext.ScenarioContext context = getCurrentScenarioContext();
+		Step step = context.getStep(testStep);
 		StartTestItemRQ rq = new StartTestItemRQ();
-		rq.setName(Utils.buildNodeName(currentScenarioContext.getStepPrefix(), step.getKeyword(), testStep.getStepText(), " "));
+		rq.setName(Utils.buildNodeName(context.getStepPrefix(), step.getKeyword(), testStep.getStepText(), " "));
 		rq.setDescription(Utils.buildMultilineArgument(testStep));
 		rq.setStartTime(Calendar.getInstance().getTime());
 		rq.setType("STEP");
@@ -78,13 +79,13 @@ public class StepReporter extends AbstractReporter {
 		rq.setCodeRef(codeRef);
 		rq.setTestCaseId(Utils.getTestCaseId(testStep, codeRef).getId());
 		rq.setAttributes(Utils.getAttributes(testStep));
-		currentStepId = RP.get().startTestItem(currentScenarioContext.getId(), rq);
+		currentStepId = launch.get().startTestItem(context.getId(), rq);
 	}
 
 	@Override
 	protected void afterStep(Result result) {
 		reportResult(result, null);
-		Utils.finishTestItem(RP.get(), currentStepId, result.getStatus());
+		Utils.finishTestItem(launch.get(), currentStepId, result.getStatus());
 		currentStepId = null;
 	}
 
@@ -95,13 +96,13 @@ public class StepReporter extends AbstractReporter {
 		rq.setStartTime(Calendar.getInstance().getTime());
 		rq.setType(isBefore ? "BEFORE_TEST" : "AFTER_TEST");
 
-		hookStepId = RP.get().startTestItem(currentScenarioContext.getId(), rq);
+		hookStepId = launch.get().startTestItem(getCurrentScenarioContext().getId(), rq);
 		hookStatus = Result.Type.PASSED;
 	}
 
 	@Override
 	protected void afterHooks(Boolean isBefore) {
-		Utils.finishTestItem(RP.get(), hookStepId, hookStatus);
+		Utils.finishTestItem(launch.get(), hookStepId, hookStatus);
 		hookStepId = null;
 	}