+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- multi-thread execution support
+- Test Case ID support
+### Fixed
+- codeRef reporting was added for every item in an item tree
 
 ## [5.0.0-BETA-12]
 ### Fixed